# mobs-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API do Projeto MOBS2",
        "description": "Documentação da API RESTful do Projeto MOBS2 para veículos e autenticação de usuários.",
        "contact": {
            "email": "seu-email@exemplo.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor de Desenvolvimento"
        }
    ],
    "paths": {
        "/register": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Registra um novo usuário no sistema",
                "description": "Cria uma nova conta de usuário com nome, email e senha e retorna um token JWT.",
                "operationId": "registerUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Victor"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "victor@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "senha123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário registrado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "type": "object"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Dados de validação inválidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Os dados fornecidos são inválidos."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Autentica um usuário e retorna um token JWT",
                "description": "Verifica as credenciais do usuário e, se corretas, gera e retorna um token JWT.",
                "operationId": "loginUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "victor@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "senha123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login bem-sucedido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/vehicles": {
            "get": {
                "tags": [
                    "Veículos"
                ],
                "summary": "Obtém a lista de todos os veículos",
                "description": "Retorna um array com todos os veículos cadastrados.",
                "operationId": "getAllVehicles",
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Vehicle"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Veículos"
                ],
                "summary": "Cria um novo veículo",
                "description": "Adiciona um novo veículo ao banco de dados e o retorna.",
                "operationId": "createVehicle",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VehicleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Veículo criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Vehicle"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Dados de validação inválidos"
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/vehicles/{vehicle_id}": {
            "get": {
                "tags": [
                    "Veículos"
                ],
                "summary": "Obtém um veículo por ID",
                "description": "Retorna um único veículo por ID.",
                "operationId": "getVehicleById",
                "parameters": [
                    {
                        "name": "vehicle_id",
                        "in": "path",
                        "description": "ID do veículo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Vehicle"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Veículo não encontrado"
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Veículos"
                ],
                "summary": "Atualiza um veículo existente",
                "description": "Atualiza os dados de um veículo por ID.",
                "operationId": "updateVehicle",
                "parameters": [
                    {
                        "name": "vehicle_id",
                        "in": "path",
                        "description": "ID do veículo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VehicleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Veículo atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Vehicle"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Veículo não encontrado"
                    },
                    "422": {
                        "description": "Dados de validação inválidos"
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Veículos"
                ],
                "summary": "Deleta um veículo",
                "description": "Remove um veículo do banco de dados por ID.",
                "operationId": "deleteVehicle",
                "parameters": [
                    {
                        "name": "vehicle_id",
                        "in": "path",
                        "description": "ID do veículo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Veículo deletado com sucesso"
                    },
                    "404": {
                        "description": "Veículo não encontrado"
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Vehicle": {
                "title": "Vehicle",
                "description": "Representação de um veículo",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "placa": {
                        "type": "string",
                        "example": "ABC-1234"
                    },
                    "modelo": {
                        "type": "string",
                        "example": "Civic"
                    },
                    "fabricante": {
                        "type": "string",
                        "example": "Honda"
                    },
                    "ano": {
                        "type": "integer",
                        "example": 2020
                    }
                },
                "type": "object"
            },
            "VehicleRequest": {
                "required": [
                    "placa",
                    "modelo",
                    "fabricante",
                    "ano"
                ],
                "properties": {
                    "placa": {
                        "type": "string",
                        "example": "ABC-1234"
                    },
                    "modelo": {
                        "type": "string",
                        "example": "Civic"
                    },
                    "fabricante": {
                        "type": "string",
                        "example": "Honda"
                    },
                    "ano": {
                        "type": "integer",
                        "example": 2020
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Insira o token JWT com o prefixo 'Bearer' no header de autorização",
                "name": "Bearer Token",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Autenticação",
            "description": "Endpoints de registro e login de usuários"
        },
        {
            "name": "Veículos",
            "description": "Endpoints para gerenciar veículos"
        }
    ]
}